First approach using local videos and passing path in endpoint is working
