+++ backend/fernec/models.py
@@ -6,3 +6,10 @@ class ImageItem(BaseModel):
 class ImagePrediction(BaseModel):
     predictions: list[float]
     emotion: str
+
+class VideoItem(BaseModel):
+    video_path: str
+    model_name: str
+
+class VideoPrediction(BaseModel):
+    prediction: list[str]
